glynnSim Replacements, Build.cmd Typo
Replaced the glynnSim variations with ports of the original C.  Turns out the thing I was debugging was somewhere else, whoops.  But they're a little more "pure" this way, so keeping.

Had a typo in the build.cmd on building the complex library for each stage.  Not even sure I ended up using this library, but there it is.